requirements.txt
requests==2.28.0
environs==9.5.0
redis==4.3.3
python-telegram-bot==13.12
